Added comments for the most doubtful feature and new constant
The most doubtful feature in my app is checking for another translation of present word in Lingualeo base, which differs from top-voted translation.
For example:
We have book-'бронировать' in Kindle database, but top-voted translation is 'книга'.
I decided to handle this situation as marking word - "is_exist".
Don't know If I was right and maybe in future I'll delete this feature.

Added MEATBALLS constant in Lingualeo for tests.
By default It's 0 but for tests it's 200. Just to check if MainWindow correctly handles
the situation, when the user has no meatballs.