Split all PyQt classes on __init__ and setVariables
When testing we need to have a link to every window/dialog to set
QTimer to its close. So, at first we just create a link to this window/widget.
Also it looks better, when ExportDialog, close_window and notif are defined
in the 'constructor' of MainWindow.
After initializing QTimer we manipulate with MainWindow by, for example,
clicking on export_button and triggering setVariable(var1, var2)